BREAKING CHANGE - rename fragment_formatters -> format
