+++ search/highlight/fragmenters/simple/simple.go
