+++ search/highlight/highlighters/simple/highlighter_simple_test.go
@@ -15,7 +15,7 @@ import (
 
 	"github.com/blevesearch/bleve/document"
 	"github.com/blevesearch/bleve/search"
-	"github.com/blevesearch/bleve/search/highlight/fragment_formatters/ansi"
+	"github.com/blevesearch/bleve/search/highlight/format/ansi"
 	sfrag "github.com/blevesearch/bleve/search/highlight/fragmenters/simple"
 )
